+++ tests/test_estimation.py
@@ -761,7 +761,11 @@ class TestBreslowFlemingHarringtonFitter:
 class TestRegressionFitters:
     @pytest.fixture
     def regression_models(self):
-        return [CoxPHFitter(), AalenAdditiveFitter(), CoxPHFitter(strata=["race", "paro", "mar", "wexp"])]
+        return [
+            CoxPHFitter(),
+            AalenAdditiveFitter(coef_penalizer=0.1),
+            CoxPHFitter(strata=["race", "paro", "mar", "wexp"]),
+        ]
 
     def test_pickle(self, rossi, regression_models):
         from pickle import dump
@@ -829,10 +833,10 @@ class TestRegressionFitters:
     def test_error_is_raised_if_using_non_numeric_data_in_fit(self, regression_models):
         df = pd.DataFrame.from_dict(
             {
-                "t": [1.0, 2.0, 3.0],
+                "t": [1.0, 2.0, 5.0],
                 "bool_": [True, True, False],
                 "int_": [1, -1, 0],
-                "uint8_": pd.Series([1, -1, 0], dtype="uint8"),
+                "uint8_": pd.Series([1, -1, 2], dtype="uint8"),
                 "string_": ["test", "a", "2.5"],
                 "float_": [1.2, -0.5, 0.0],
                 "categorya_": pd.Series([1, 2, 3], dtype="category"),
@@ -840,12 +844,12 @@ class TestRegressionFitters:
             }
         )
 
-        for fitter in [CoxPHFitter(), AalenAdditiveFitter()]:
+        for fitter in [CoxPHFitter(), AalenAdditiveFitter(coef_penalizer=0.01)]:
             for subset in [["t", "categorya_"], ["t", "categoryb_"], ["t", "string_"]]:
                 with pytest.raises(TypeError):
                     fitter.fit(df[subset], duration_col="t")
 
-            for subset in [["t", "bool_"], ["t", "int_"], ["t", "float_"], ["t", "uint8_"]]:
+            for subset in [["t", "uint8_"]]:
                 fitter.fit(df[subset], duration_col="t")
 
     def test_regression_model_has_score_(self, regression_models, rossi):
@@ -854,6 +858,17 @@ class TestRegressionFitters:
             assert not hasattr(fitter, "score_")
             fitter.fit(rossi, duration_col="week", event_col="arrest")
             assert hasattr(fitter, "score_")
+
+    def test_regression_model_updates_score_(self, regression_models, rossi):
+
+        for fitter in regression_models:
+            assert not hasattr(fitter, "score_")
+            fitter.fit(rossi, duration_col="week", event_col="arrest")
+            assert hasattr(fitter, "score_")
+            first_score_ = fitter.score_
+
+            fitter.fit(rossi.head(20), duration_col="week", event_col="arrest")
+            assert first_score_ != fitter.score_
 
     def test_error_is_thrown_if_there_is_nans_in_the_duration_col(self, regression_models, rossi):
         rossi.loc[3, "week"] = None
@@ -872,6 +887,21 @@ class TestCoxPHFitter:
     @pytest.fixture
     def cph(self):
         return CoxPHFitter()
+
+    def test_cph_doesnt_modify_original_dataframe(self, cph):
+        df = pd.DataFrame(
+            {
+                "var1": [-0.71163379, -0.87481227, 0.99557251, -0.83649751, 1.42737105],
+                "T": [5, 6, 7, 8, 9],
+                "E": [1, 1, 1, 1, 1],
+                "W": [1, 1, 1, 1, 1],
+            }
+        )
+
+        cph.fit(df, "T", "E", weights_col="W")
+        assert df.dtypes["E"] == int
+        assert df.dtypes["W"] == int
+        assert df.dtypes["T"] == int
 
     def test_cph_will_handle_times_with_only_censored_individuals(self, rossi):
         rossi_29 = rossi.iloc[0:10].copy()
@@ -2327,35 +2357,20 @@ Likelihood ratio test = 33.27 on 7 df, log(p)=-10.65
 
 
 class TestAalenAdditiveFitter:
-    def test_nn_cumulative_hazard_will_set_cum_hazards_to_0(self, rossi):
-        aaf = AalenAdditiveFitter(nn_cumulative_hazard=False)
-        aaf.fit(rossi, event_col="arrest", duration_col="week")
-        cum_hazards = aaf.predict_cumulative_hazard(rossi)
-        assert (cum_hazards < 0).stack().mean() > 0
+    @pytest.fixture()
+    def aaf(self):
+        return AalenAdditiveFitter()
 
-        aaf = AalenAdditiveFitter(nn_cumulative_hazard=True)
-        aaf.fit(rossi, event_col="arrest", duration_col="week")
-        cum_hazards = aaf.predict_cumulative_hazard(rossi)
-        assert (cum_hazards < 0).stack().mean() == 0
-
-    def test_using_a_custom_timeline_in_static_fitting(self, rossi):
-        aaf = AalenAdditiveFitter()
-        timeline = np.arange(10)
-        aaf.fit(rossi, event_col="arrest", duration_col="week", timeline=timeline)
-        npt.assert_array_equal(aaf.hazards_.index.values, timeline)
-        npt.assert_array_equal(aaf.cumulative_hazards_.index.values, timeline)
-        npt.assert_array_equal(aaf.variance_.index.values, timeline)
-        npt.assert_array_equal(aaf.timeline, timeline)
-
-    def test_using_a_custom_timeline_in_varying_fitting(self):
-        panel_dataset = load_panel_test()
-        aaf = AalenAdditiveFitter()
-        timeline = np.arange(10)
-        aaf.fit(panel_dataset, id_col="id", duration_col="t", timeline=timeline)
-        npt.assert_array_equal(aaf.hazards_.index.values, timeline)
-        npt.assert_array_equal(aaf.cumulative_hazards_.index.values, timeline)
-        npt.assert_array_equal(aaf.variance_.index.values, timeline)
-        npt.assert_array_equal(aaf.timeline, timeline)
+    def test_slope_tests_against_R(self, aaf, regression_dataset):
+        """
+        df['E'] = 1
+        a = aareg(formula=Surv(T, E) ~ var1 + var2 + var3, data=df)
+        plot(a)
+        summary(a, test='nrisk')
+        """
+        regression_dataset["E"] = 1
+        aaf.fit(regression_dataset, "T", "E")
+        npt.assert_allclose(aaf.summary["slope(coef)"], [0.05141401, 0.01059746, 0.03923360, 0.07753566])
 
     def test_penalizer_reduces_norm_of_hazards(self, rossi):
         from numpy.linalg import norm
@@ -2400,16 +2415,8 @@ class TestAalenAdditiveFitter:
         X = pd.DataFrame(np.random.randn(n, d))
         T = np.random.exponential(size=n)
         X["T"] = T
-        aaf = AalenAdditiveFitter()
+        aaf = AalenAdditiveFitter(coef_penalizer=0.01)
         aaf.fit(X, duration_col="T")
-
-    def test_aaf_panel_dataset_with_no_censorship(self):
-        panel_dataset = load_panel_test()
-        aaf = AalenAdditiveFitter()
-        aaf.fit(panel_dataset, id_col="id", duration_col="t")
-        expected = pd.Series([True] * 9, index=range(1, 10))
-        expected.index.name = "id"
-        assert_series_equal(aaf.event_observed, expected)
 
     def test_aalen_additive_median_predictions_split_data(self):
         # This tests to make sure that my median predictions statisfy
@@ -2420,10 +2427,13 @@ class TestAalenAdditiveFitter:
         timeline = np.linspace(0, 70, 5000)
         hz, coef, X = generate_hazard_rates(n, d, timeline)
         T = generate_random_lifetimes(hz, timeline)
+
         X["T"] = T
         X = X.replace([np.inf, -np.inf], 10.0)
+        # del X[5]
+
         # fit it to Aalen's model
-        aaf = AalenAdditiveFitter()
+        aaf = AalenAdditiveFitter(coef_penalizer=0.5, fit_intercept=False)
         aaf.fit(X, "T")
 
         # predictions
@@ -2431,7 +2441,7 @@ class TestAalenAdditiveFitter:
         assert abs((T_pred.values > T).mean() - 0.5) < 0.05
 
     def test_dataframe_input_with_nonstandard_index(self):
-        aaf = AalenAdditiveFitter()
+        aaf = AalenAdditiveFitter(coef_penalizer=0.1)
         df = pd.DataFrame(
             [(16, True, True), (1, True, True), (4, False, True)],
             columns=["duration", "done_feeding", "white"],
@@ -2440,7 +2450,7 @@ class TestAalenAdditiveFitter:
         aaf.fit(df, duration_col="duration", event_col="done_feeding")
 
     def test_crossval_for_aalen_add(self, data_pred2, data_pred1):
-        aaf = AalenAdditiveFitter()
+        aaf = AalenAdditiveFitter(coef_penalizer=0.1)
         for data_pred in [data_pred1, data_pred2]:
             mean_scores = []
             for repeat in range(20):
@@ -2452,12 +2462,52 @@ class TestAalenAdditiveFitter:
             assert np.mean(mean_scores) > expected, msg.format(expected, np.mean(scores))
 
     def test_predict_cumulative_hazard_inputs(self, data_pred1):
-        aaf = AalenAdditiveFitter()
+        aaf = AalenAdditiveFitter(coef_penalizer=0.001)
         aaf.fit(data_pred1, duration_col="t", event_col="E")
         x = data_pred1.iloc[:5].drop(["t", "E"], axis=1)
         y_df = aaf.predict_cumulative_hazard(x)
         y_np = aaf.predict_cumulative_hazard(x.values)
         assert_frame_equal(y_df, y_np)
+
+    def test_aalen_additive_fitter_versus_R(self, aaf, rossi):
+        """
+        a = aareg(formula=Surv(week, arrest) ~ fin + age + race+ wexp + mar + paro + prio, data=head(rossi, 432))
+        """
+        aaf.fit(rossi, "week", "arrest")
+        actual = aaf.hazards_
+        npt.assert_allclose(actual.loc[:2, "fin"].tolist(), [-0.004628582, -0.005842295], rtol=1e-06)
+        npt.assert_allclose(actual.loc[:2, "prio"].tolist(), [-1.268344e-03, 1.119377e-04], rtol=1e-06)
+        npt.assert_allclose(actual.loc[:2, "baseline"].tolist(), [1.913901e-02, -3.297233e-02], rtol=1e-06)
+
+    def test_aalen_additive_fitter_versus_R_with_weights(self, aaf, regression_dataset):
+        """
+        df['E'] = 1
+        a = aareg(formula=Surv(T, E) ~ var1 + var2, data=df, weights=var3)
+        a$coefficient
+        """
+        regression_dataset["E"] = 1
+        aaf.fit(regression_dataset, "T", "E", weights_col="var3")
+        actual = aaf.hazards_
+        npt.assert_allclose(actual.iloc[:3]["var1"].tolist(), [1.301523e-02, -4.925302e-04, 2.304792e-02], rtol=1e-06)
+        npt.assert_allclose(
+            actual.iloc[:3]["baseline"].tolist(), [-9.672957e-03, 1.439187e-03, 1.838915e-03], rtol=1e-06
+        )
+
+    def test_cumulative_hazards_versus_R(self, aaf, regression_dataset):
+        """
+        df['E'] = 1
+        a = aareg(formula=Surv(T, E) ~ var1 + var2 + var3, data=df)
+        c = a$coefficient
+        apply(c, 2, cumsum)
+        """
+        regression_dataset["E"] = 1
+
+        aaf.fit(regression_dataset, "T", "E")
+        actual = aaf.cumulative_hazards_.loc[regression_dataset["T"].max()]
+        npt.assert_allclose(actual["baseline"], 2.1675130235, rtol=1e-06)
+        npt.assert_allclose(actual["var1"], 0.6820086125, rtol=1e-06)
+        npt.assert_allclose(actual["var2"], -0.0776583514, rtol=1e-06)
+        npt.assert_allclose(actual["var3"], 0.5515174017, rtol=1e-06)
 
 
 class TestCoxTimeVaryingFitter: